+++ cookbooks/reboot-handler/spec/spec_helper.rb
@@ -1,0 +1,6 @@
+# encoding: UTF-8
+
+require 'chef'
+require 'chefspec'
+require 'chefspec/berkshelf'
+require 'chefspec/deprecations'