+++ cookbooks/yasm/libraries/helpers.rb
@@ -2,22 +2,29 @@
 # Cookbook Name:: yasm
 # Library:: helpers
 #
-# Copyright 2012-2013, Escape Studios
+# Copyright 2012-2014, Escape Studios
 #
 
+# The YASM module name-spaces all the classes of the YASM-cookbook
+#
 module YASM
-    module Helpers
-        #Returns an array of package names that will install YASM on a node.
-        #Package names returned are determined by the platform running this recipe.
-        def yasm_packages
-            value_for_platform(
-                [ "ubuntu" ] => { "default" => [ "yasm" ] },
-                "default" => [ "yasm" ]
-            )
-        end
+  # Specific helpers
+  #
+  module Helpers
+    # returns an array of package names that will install YASM on a node
+    # package names returned are determined by the platform running this recipe.
+    def yasm_packages
+      value_for_platform(
+        ['ubuntu'] => { 'default' => ['yasm'] },
+        'default' => ['yasm']
+      )
     end
+  end
 end
 
-class Chef::Recipe
+class Chef
+  # ==Chef::Recipe
+  class Recipe
     include YASM::Helpers
+  end
 end